pyproject.toml
[tool.poetry]
name = "travel-site"
version = "0.1.0"
description = ""
authors = ["alone"]

[virtualenvs]
create = true
in-project = true


[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.9"
flake8 = "^4.0.1"
pep8 = "^1.7.1"
pytest = "^6.2.5"
django-crispy-forms = "^1.13.0"
django-environ = "^0.8.1"
loguru = "^0.5.3"
Pillow = "^8.4.0"
django-phonenumber-field = "^6.0.0"
phonenumbers = "^8.12.37"
gunicorn = "^20.1.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

